Read hyperparameters from the file Hyperparameters.json in the model directory if it exists

+++ src/models/morphon/__init__.py
@@ -16,14 +16,19 @@ def make_model(
     model_directory: str,
     args: Optional[dict[str, Any]] = None):
     hyperparameters_directory = model_subtype + '-pos' if use_features else model_subtype
-    with open(path.join('default_hyperparameters', model_type, hyperparameters_directory, 'Hyperparameters.json')) as fin:
+    hyperparameters_file = path.join(model_directory, 'Hyperparameters.json')
+    if path.exists(hyperparameters_file):
+      with open(hyperparameters_file, 'r', encoding='utf-8') as fin:
+        hyperparameters = json.load(fin)
+    else:
+      with open(path.join('default_hyperparameters', model_type, hyperparameters_directory, 'Hyperparameters.json')) as fin:
         default_hyperparameters = json.load(fin)
-    hyperparameters = default_hyperparameters if args is None else \
-      {key: args[key] if key in args and args[key] is not None
-                      else default_hyperparameters[key]
-       for key in default_hyperparameters}
-    with open(path.join(model_directory, 'Hyperparameters.json'), 'w', encoding='utf-8') as fout:
-      json.dump(hyperparameters, fout)
+      hyperparameters = default_hyperparameters if args is None else \
+        {key: args[key] if key in args and args[key] is not None
+                        else default_hyperparameters[key]
+        for key in default_hyperparameters}
+      with open(hyperparameters_file, 'w', encoding='utf-8') as fout:
+        json.dump(hyperparameters, fout)
     match model_type:
         case 'tagger':
             return make_tagger(model_subtype, vocabularies, hyperparameters, device)
